refactor CRs communication
